course deletion implemented and tested

+++ src/main/java/com/revature/p0/screens/FacultyRegistrationScreen.java
@@ -4,6 +4,7 @@ import com.revature.p0.documents.AppUser;
 import com.revature.p0.documents.Course;
 import com.revature.p0.questions.NavigateScreenQuestion;
 import com.revature.p0.questions.Question;
+import com.revature.p0.questions.courseQuestions.CourseTagQuestion;
 import com.revature.p0.services.CourseService;
 import com.revature.p0.services.UserService;
 import com.revature.p0.util.QuestionFactory;
@@ -32,13 +33,13 @@ public class FacultyRegistrationScreen extends Screen{
         String menu = "You're on the course management screen.\n" +
                 "1)Create a course\n" +
                 "2)Remove a course\n" +
-                "3)Update a course\n" +
+                "3)Update course info\n" +
                 "4)Back\n";
 
         System.out.println(courseList);
         System.out.println(menu);
 
-        Question prompt = new NavigateScreenQuestion(3);
+        Question prompt = new NavigateScreenQuestion(4);
         String userInput = reader.readLine();
         while (!prompt.validAnswer(userInput)) {
             userInput = reader.readLine();
@@ -63,13 +64,16 @@ public class FacultyRegistrationScreen extends Screen{
                 System.out.println("Registration successful!");
                 break;
             case "2":
-                qFactory.getCourseQuestion("coursetag",service);
+                //Forced solution. Overloaded question validation method since finding an existing course IS a valid answer, here.
+                CourseTagQuestion question = (CourseTagQuestion) qFactory.getCourseQuestion("coursetag", service);
                 String tag = reader.readLine();
-                service.removeCourse(tag);
+                if (question.validAnswer(tag, "delete"))
+                    service.removeCourse(tag);
                 break;
-            case "3:":
+            case "3":
+                
                 break;
-            case "4:":
+            case "4":
                 router.previousScreen();
         }
     }
